// src/app/models/te/tecapr.ts
export class teapert{
    apc_nume: string;
    apc_desc: string;
    apc_cont: number;
}

export class teasist{
    emp_codi:   number;
    ter_coda:   string;
    apc_cont:   number;
    asi_desc:   string;
}

export class tecaper {
    asi_fech: string;
    apc_nume: string;
    cac_codi: string;
    cac_nomb: string;
    ter_coda: string;
    ter_noco: string;
    cli_coda: string;
    cli_noco: string;
    das_asis: string;
    asi_tipo: string;
}