// src/app/models/te/teasist.ts
export class teapert{    
    apc_cont: number;
    apc_nume: number;
    cut_codi: number;
    cut_nomb: string;
}

export class teasist{
    constructor(){}
    emp_codi: number;
    ter_coda: string;
    apc_cont: number;
    asi_desc: string;
}

export class teasise{
    emp_codi: number;
    asi_cont: number;
    cli_coda: string;
    cli_noco: string;
    das_asis: boolean;
}